requirements.txt
requests
beautifulsoup4
tqdm
python-dotenv